requirements.txt
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
idna==3.11
MarkupSafe==3.0.3
pycparser==3.0
python-dotenv==1.2.1
python-http-client==3.3.7
requests==2.32.5
sendgrid==6.12.5
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5

# Trading bot dependencies
python-binance==1.0.19
pandas==2.2.0
numpy==1.26.4
